A;adir dropdown de cancha o ubicacion
